+++ CHANGES.rst
@@ -7,6 +7,10 @@
 
 * Convert cmdline argument parser from docopt to argparse
 
+* Add configuration file support
+
+* Refactor logging/ui
+
 1.5.0
 =====